Refacto to another file + add icon in card for the type
